clear out unused util, covered by qcel, or updated in psi4
