+++ qcdb/util/__init__.py
@@ -1,5 +1,4 @@
 from .mpl import plot_coord
-from .misc import update_with_error, filter_comments
 from .regex import yes, no, der0th, der1st, der2nd, parse_dertype
 from .paths import search_file, import_ignorecase
 from .text import banner, find_approximate_string_matches
