+++ qcdb/util/misc.py
@@ -1,48 +1,5 @@
-import re
-import sys
-import math
-
-import numpy as np
-
 import qcelemental as qcel
 
-def update_with_error(a, b, path=None):
-    """Merges `b` into `a` like dict.update; however, raises KeyError if values of a
-    key shared by `a` and `b` conflict.
-
-    Adapted from: https://stackoverflow.com/a/7205107
-
-    """
-    if path is None:
-        path = []
-    for key in b:
-        if key in a:
-            if isinstance(a[key], dict) and isinstance(b[key], dict):
-                update_with_error(a[key], b[key], path + [str(key)])
-            elif a[key] == b[key]:
-                pass  # same leaf value
-            elif a[key] is None:
-                a[key] = b[key]
-            elif (isinstance(a[key], (list, tuple)) and
-                  not isinstance(a[key], str) and
-                  isinstance(b[key], (list, tuple)) and
-                  not isinstance(b[key], str) and
-                  len(a[key]) == len(b[key]) and
-                  all((av is None or av == bv) for av, bv in zip(a[key], b[key]))):  # yapf: disable
-                a[key] = b[key]
-            else:
-                raise KeyError('Conflict at {}: {} vs. {}'.format('.'.join(path + [str(key)]), a[key], b[key]))
-        else:
-            a[key] = b[key]
-    return a
-
-
-def filter_comments(string):
-    """Remove from `string` any Python-style comments ('#' to end of line)."""
-
-    comment = re.compile(r'(^|[^\\])#.*')
-    string = re.sub(comment, '', string)
-    return string
 
 
 def process_units(molrec):
@@ -82,7 +39,7 @@ def process_units(molrec):
     elif units in ['Angstrom', 'Bohr'] and input_units_to_au is not None:
         expected_iutau = a2b if units == 'Angstrom' else 1.
 
-        if perturn_check(input_units_to_au, expected_iutau):
+        if perturb_check(input_units_to_au, expected_iutau):
             funits = units
             fiutau = input_units_to_au
         else:
